+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Starting to properly keep track of changes in this repository
 ## [Unreleased]
 ### Added
 - `spiro_analysis.preprocess.merge_strobe_triplet` function added
+- `spiro_analysis.preprocess.merge_strobe_directory` function added
 
 ### Changed
 - Got rid of the imports in the `spiro_analysis/__init__.py` file for now